AppWeb_alertas/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
yfinance
gspread
oauth2client==4.1.3
openpyxl